# UniQUE-API/src/docs/swagger.yaml
basePath: /
definitions:
...
  routes.ApplicationDTO:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      privacy_policy_url:
        type: string
      public_client:
        type: boolean
      updated_at:
        type: string
      user_id:
        type: string
      website_url:
        type: string
    type: object
  routes.ApplicationListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/routes.ApplicationDTO'
        type: array
    type: object
...
  routes.CreateApplicationRequest:
    properties:
      client_secret:
        type: string
      description:
        type: string
      name:
        type: string
      privacy_policy_url:
        type: string
      public_client:
        type: boolean
      user_id:
        type: string
      website_url:
        type: string
    required:
    - name
    - user_id
    type: object
...
  routes.CreateRedirectURIRequest:
    properties:
      uri:
        type: string
    required:
    - uri
    type: object
...
  routes.Nullable-bool:
    properties:
      set:
        type: boolean
      value:
        type: boolean
    type: object
  routes.Nullable-string:
    properties:
      set:
        type: boolean
      value:
        type: string
    type: object
...
  routes.PatchApplicationRequest:
    properties:
      client_secret:
        $ref: '#/definitions/routes.Nullable-string'
      description:
        $ref: '#/definitions/routes.Nullable-string'
      name:
        $ref: '#/definitions/routes.Nullable-string'
      privacy_policy_url:
        $ref: '#/definitions/routes.Nullable-string'
      public_client:
        $ref: '#/definitions/routes.Nullable-bool'
      website_url:
        $ref: '#/definitions/routes.Nullable-string'
    type: object
...
  routes.RedirectURIDTO:
    properties:
      created_at:
        type: string
      updated_at:
        type: string
      uri:
        type: string
    type: object
  routes.RedirectURIListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/routes.RedirectURIDTO'
        type: array
    type: object
...
  routes.UpdateApplicationRequest:
    properties:
      client_secret:
        $ref: '#/definitions/routes.Nullable-string'
      description:
        $ref: '#/definitions/routes.Nullable-string'
      name:
        $ref: '#/definitions/routes.Nullable-string'
      privacy_policy_url:
        $ref: '#/definitions/routes.Nullable-string'
      public_client:
        $ref: '#/definitions/routes.Nullable-bool'
      website_url:
        $ref: '#/definitions/routes.Nullable-string'
    type: object
...
info:
  contact: {}
paths:
...
  /applications:
    get:
      description: List third-party applications
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ApplicationListResponse'
      summary: List applications
      tags:
      - applications
    post:
      consumes:
      - application/json
      description: Register a new third-party application
      parameters:
      - description: Create application
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/routes.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.ApplicationDTO'
      summary: Create an application
      tags:
      - applications
  /applications/{id}:
    get:
      description: Get a single application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ApplicationDTO'
      summary: Get an application
      tags:
      - applications
    patch:
      consumes:
      - application/json
      description: パッチ更新。指定されたフィールドのみ更新（null を送ると NULL に）
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Patch application
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/routes.PatchApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ApplicationDTO'
      summary: Partially update an application
      tags:
      - applications
    put:
      consumes:
      - application/json
      description: Update application fields
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Update application
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/routes.UpdateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ApplicationDTO'
      summary: Update an application
      tags:
      - applications
  /applications/{id}/redirect_uris:
    delete:
      description: Delete a registered redirect URI by application id and uri (query
        param)
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Redirect URI
        in: query
        name: uri
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete redirect URI for an application
      tags:
      - applications
    get:
      description: Get redirect URIs registered for a given application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.RedirectURIListResponse'
      summary: List redirect URIs for an application
      tags:
      - applications
    post:
      consumes:
      - application/json
      description: Register a new redirect URI for the application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Create redirect URI
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.CreateRedirectURIRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.RedirectURIDTO'
      summary: Create redirect URI for an application
      tags:
      - applications
...
  /users/{id}/apps:
    get:
      description: Get applications owned by a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ApplicationListResponse'
      summary: List applications for a user
      tags:
      - users
...
swagger: "2.0"